requirements.txt
jupyter
pandas
pyarrow
scikit-learn
numpy
transformers
matplotlib
langdetect
setfit
torcheval
sentencepiece
flair